+++ eventbus-core/build.gradle
@@ -18,7 +18,6 @@ dependencies {
   compile("javax.validation:validation-api:1.1.0.Final")
 
   testCompile("junit:junit:4.11")
-  testCompile("de.bechte.junit:junit-hierarchicalcontextrunner:4.11.3")
   testCompile("ch.qos.logback:logback-classic:1.1.3")
   testCompile("nl.jqno.equalsverifier:equalsverifier:1.7.5")
 }
