+++ eventbus-core/src/main/java/eu/tripledframework/eventbus/internal/infrastructure/interceptor/LoggingEventBusInterceptor.java
@@ -17,6 +17,7 @@ package eu.tripledframework.eventbus.internal.infrastructure.interceptor;
 
 import eu.tripledframework.eventbus.EventBusInterceptor;
 import eu.tripledframework.eventbus.internal.domain.InterceptorChain;
+import eu.tripledframework.eventbus.internal.domain.UnitOfWork;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -28,7 +29,7 @@ public class LoggingEventBusInterceptor implements EventBusInterceptor {
   private static final Logger LOGGER = LoggerFactory.getLogger(LoggingEventBusInterceptor.class);
 
   @Override
-  public <ReturnType> ReturnType intercept(InterceptorChain<ReturnType> chain, Object event) {
+  public <ReturnType> ReturnType intercept(InterceptorChain<ReturnType> chain, Object event, UnitOfWork unitOfWork) {
     LOGGER.debug("Executing command {}", event.getClass().getSimpleName());
     Instant start = Instant.now();
     try {
