--- eventbus-core/src/main/java/eu/tripledframework/eventbus/internal/infrastructure/invoker/DuplicateInvokerRegistrationException.java
@@ -1,7 +1,0 @@
-package eu.tripledframework.eventbus.internal.infrastructure.invoker;
-
-public class DuplicateInvokerRegistrationException extends RuntimeException {
-  public DuplicateInvokerRegistrationException(String message) {
-    super(message);
-  }
-}
