+++ tripledframework-demo/build.gradle
@@ -37,11 +37,11 @@ dependencies {
   compile project(":spring-boot-eventbus-starter")
   compile project(":eventstore-core")
   compile("org.springframework.boot:spring-boot-starter-web")
+  compile("org.springframework.boot:spring-boot-starter-security")
   compile("org.hibernate:hibernate-validator")
   runtime("org.hsqldb:hsqldb")
-  // commented to resolve ELExpressionFactory class not found when launching the JAR.
-  //providedRuntime("org.springframework.boot:spring-boot-starter-tomcat")
 
   testCompile("org.springframework.boot:spring-boot-starter-test")
+  testCompile("org.springframework.security:spring-security-test")
   testCompile("com.jayway.jsonpath:json-path")
 }
