+++ tripledframework-demo/src/main/java/eu/tripledframework/demo/infrastructure/InMemoryHelloEventStore.java
@@ -25,7 +25,7 @@ import org.springframework.stereotype.Component;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import eu.tripledframework.demo.SaidHelloDomainEvent;
+import eu.tripledframework.demo.model.SaidHelloDomainEvent;
 
 @Handler
 @Component
