+++ tripledframework-demo/src/main/java/eu/tripledframework/demo/model/SaidHelloDomainEvent.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.tripledframework.demo;
+package eu.tripledframework.demo.model;
 
 public class SaidHelloDomainEvent {
