+++ tripledframework-demo/src/main/java/eu/tripledframework/demo/security/SpringSecurityAwareUnitOfWorkFactory.java
@@ -1,0 +1,11 @@
+package eu.tripledframework.demo.security;
+
+import eu.tripledframework.eventbus.internal.domain.UnitOfWork;
+import eu.tripledframework.eventbus.internal.domain.UnitOfWorkFactory;
+
+public class SpringSecurityAwareUnitOfWorkFactory implements UnitOfWorkFactory {
+  @Override
+  public UnitOfWork create() {
+    return new SpringSecurityAwareUnitOfWork();
+  }
+}
